separate core from addon
