+++ .github/workflows/checks.yml
@@ -1,11 +1,11 @@
 name: checks
- 
+
 on:
   push:
-    branches: 
+    branches:
       - "*"
   pull_request:
-    branches: 
+    branches:
       - "*"
 
 jobs:
@@ -13,9 +13,9 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-        with: 
-         submodules: recursive
-      
+        with:
+          submodules: recursive
+
       - name: Setup Python
         uses: actions/setup-python@v4
         with:
@@ -25,4 +25,4 @@ jobs:
         run: python -m pip install aqt mypy pyqt5-stubs types-requests types-simplejson pycryptodome
 
       - name: Run mypy
-        run: python -m mypy ./src/addon
+        run: python -m mypy ./src
