+++ .github/workflows/create_release.yml
@@ -12,35 +12,35 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-        with: 
+        with:
           submodules: recursive
-      
+
       - name: Setup Python
         uses: actions/setup-python@v4
         with:
           python-version: 3.8
-      
+
       - name: Install dependencies
         run: python -m pip install simplejson
-      
+
       - name: Bundle dependencies
         run: pip install --no-deps --target=./src/addon/libs -r requirements.txt
-          
+
       - name: Run release/new_version.py
         run: python ./release/new_version.py ${{ github.event.inputs.version }}
-      
+
       - name: Commit changes to git
         uses: EndBug/add-and-commit@v7.1.1
         with:
           message: Bump Version to v${{ github.event.inputs.version }}
           tag: ${{ github.event.inputs.version }}
-      
+
       - name: Bundle google api key
         run: python ./release/api_obsfucate.py "${{ secrets.GOOGLE_API_KEY }}"
-          
+
       - name: Create ankiaddon file
-        run: cd src/addon ; zip -r ../../media_import_v${{ github.event.inputs.version }}.ankiaddon * ; cd ../../
-      
+        run: cd src ; zip -r ../media_import_v${{ github.event.inputs.version }}.ankiaddon * ; cd ../
+
       - name: Create github release and upload ankiaddon file
         uses: svenstaro/upload-release-action@2.2.1
         with:
