+++ src/__init__.py
@@ -1,0 +1,4 @@
+from .addon import open_import_dialog
+from .anking import setupMenu
+
+setupMenu(open_import_dialog)
